Intergrate time into actions layer
