+++ src/arena/views.rs
@@ -28,7 +28,7 @@ pub trait View {
 
 pub use debug::DebugView;
 pub use infobar::InfoBarView;
-pub use log::{LogComponent, LogView};
+pub use log::LogView;
 pub use map::{screen_rect_for_map_grid, screen_to_map_position, MapView, MAP_CORNER_Y, TILE_SIZE};
 pub use skillbar::SkillBarView;
